Fix plugin test
Installer constructor fails because current composer checks and tries to
remove disabled installers but ends up calling a method on the null
returned by the mock. This mocks up the new method call though this
could be prone to failures on future versions. The dangers of mocking
things you don't own.